fix:migrate leave absent

+++ database/migrations/2024_09_17_052159_add_total_days_to_leave_requests_table.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('leave_requests', function (Blueprint $table) {
-            $table->integer('total_days')->unsigned();
+            if (Schema::hasColumn('leave_requests', 'total_days')) {
+                $table->integer('total_days')->unsigned();
+            }
             if (Schema::hasColumn('leave_requests', 'hrd_id')) {
                 $table->foreignId('hrd_id')->nullable()->constrained('employees')->nullOnDelete()->cascadeOnUpdate();
             }
@@ -28,15 +30,15 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('leave_requests', function (Blueprint $table) {
-            $table->dropColumn('total_days');
-            Schema::table('leave_requests', function (Blueprint $table) {
-                if (!Schema::hasColumn('leave_requests', 'hrd_id')) {
-                    $table->unsignedBigInteger('hrd_id')->nullable();
-                }
-                if (!Schema::hasColumn('leave_requests', 'hrd_approved_at')) {
-                    $table->timestamp('hrd_approved_at')->nullable();
-                }
-            });
+            if (!Schema::hasColumn('leave_requests', 'total_days')) {
+                $table->dropColumn('total_days');
+            }
+            if (!Schema::hasColumn('leave_requests', 'hrd_id')) {
+                $table->unsignedBigInteger('hrd_id')->nullable();
+            }
+            if (!Schema::hasColumn('leave_requests', 'hrd_approved_at')) {
+                $table->timestamp('hrd_approved_at')->nullable();
+            }
         });
     }
 };
